165:FixingtheErrorHandler (to see the actual error message)
Fixing the Error Handler
Section 7, Lecture 165
Unfortunately, the error handler won't correctly in its current state.

Right now, we try to extract the error message like this:

.subscribe(
  ...,
  error => {
    loading.dismiss();
    this.handleError(error.message);
  }
);
But this actually has two problems:

error  still is the original Response  object - we first have to extract the body  (and convert it to a JS object) via json() .
The actual message is not stored in a property named message  but in a property named error .
The correct code therefore looks like this:

.subscribe(
  ...
  ,
  error => {
    loading.dismiss();
    this.handleError(error.json().error);
  }
);
Replace the code in the shopping-list.ts and the recipes.ts file and you should see the actual error messages.